add new class CVehicleAnalysis

+++ Homework_02_03_2019/001ConAppVehicle/CCar.cs
@@ -8,8 +8,9 @@ namespace _001ConAppVehicle
 {
     class CCar : CVehicle
     {
-        public CCar(string name, int x, int y, int price, int speed, int yearIssue) : base(name, x, y, price, speed, yearIssue)
+        public CCar(string name, int x, int y, int price, double speed, int yearIssue) : base(name, x, y, price, speed, yearIssue)
         {
         }
+
     }
 }
